+++ flutter_searchbox/lib/src/components/searchwidgetconnector.dart
@@ -220,6 +220,26 @@ class _SearchWidgetListener<S, ViewModel> extends StatefulWidget {
 
   final Map? distinctFieldConfig;
 
+  /// This prop is used to set the timeout value for HTTP requests.
+  /// Defaults to 30 seconds.
+  Duration httpRequestTimeout;
+
+  /// Configure whether the DSL query is generated with the compound clause of [CompoundClauseType.must] or [CompoundClauseType.filter]. If nothing is passed the default is to use [CompoundClauseType.must].
+  /// Setting the compound clause to filter allows search engine to cache and allows for higher throughput in cases where scoring isn’t relevant (e.g. term, geo or range type of queries that act as filters on the data)
+  ///
+  /// This property only has an effect when the search engine is either elasticsearch or opensearch.
+  /// > Note: `compoundClause` is supported with v8.16.0 (server) as well as with serverless search.
+  ///
+  ///   ///
+  /// For example,
+  /// ```dart
+  /// SearchBox(
+  ///   ...
+  ///   compoundClause:  CompoundClauseType.filter
+  /// )
+  /// ```
+  CompoundClauseType? compoundClause;
+
   /* ---- callbacks to create the side effects while querying ----- */
 
   final Future Function(dynamic value)? beforeValueChange;
@@ -308,8 +328,9 @@ class _SearchWidgetListener<S, ViewModel> extends StatefulWidget {
     this.value,
     this.distinctField,
     this.distinctFieldConfig,
-  }) : super(key: key);
-
+    this.httpRequestTimeout = const Duration(seconds: 30),
+    this.compoundClause,
+  }) : super(key: key) {}
   @override
   _SearchWidgetListenerState createState() =>
       _SearchWidgetListenerState<S, ViewModel>(
@@ -370,6 +391,8 @@ class _SearchWidgetListener<S, ViewModel> extends StatefulWidget {
           'value': value,
           'distinctField': distinctField,
           'distinctFieldConfig': distinctFieldConfig,
+          'httpRequestTimeout': httpRequestTimeout,
+          'compoundClause': compoundClause
         },
         builder: builder,
         subscribeTo: subscribeTo,
@@ -810,6 +833,26 @@ class SearchWidgetConnector<S, ViewModel> extends StatelessWidget {
   /// ```
   final Map? distinctFieldConfig;
 
+  /// This prop is used to set the timeout value for HTTP requests.
+  /// Defaults to 30 seconds.
+  Duration httpRequestTimeout;
+
+  /// Configure whether the DSL query is generated with the compound clause of [CompoundClauseType.must] or [CompoundClauseType.filter]. If nothing is passed the default is to use [CompoundClauseType.must].
+  /// Setting the compound clause to filter allows search engine to cache and allows for higher throughput in cases where scoring isn’t relevant (e.g. term, geo or range type of queries that act as filters on the data)
+  ///
+  /// This property only has an effect when the search engine is either elasticsearch or opensearch.
+  /// > Note: `compoundClause` is supported with v8.16.0 (server) as well as with serverless search.
+  ///
+  ///   ///
+  /// For example,
+  /// ```dart
+  /// SearchBox(
+  ///   ...
+  ///   compoundClause:  CompoundClauseType.filter
+  /// )
+  /// ```
+  CompoundClauseType? compoundClause;
+
   /* ---- callbacks to create the side effects while querying ----- */
 
   /// It is a callback function which accepts component's future **value** as a
@@ -920,10 +963,13 @@ class SearchWidgetConnector<S, ViewModel> extends StatelessWidget {
     this.value,
     this.distinctField,
     this.distinctFieldConfig,
+    this.httpRequestTimeout = const Duration(seconds: 30),
+    this.compoundClause,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    print('L939 httpRequestTimeout $httpRequestTimeout');
     return _SearchBaseConnector(
         child: (searchbase) => _SearchWidgetListener(
               id: id,
@@ -987,6 +1033,8 @@ class SearchWidgetConnector<S, ViewModel> extends StatelessWidget {
               value: value,
               distinctField: distinctField,
               distinctFieldConfig: distinctFieldConfig,
+              httpRequestTimeout: httpRequestTimeout,
+              compoundClause: compoundClause,
             ));
   }
 }