+++ flutter_searchbox_ui/lib/src/components/reactive_google_map.dart
@@ -617,6 +617,8 @@ class _ReactiveGoogleMapState extends State<ReactiveGoogleMap> {
       onError: widget.onError,
       onRequestStatusChange: widget.onRequestStatusChange,
       onQueryChange: widget.onQueryChange,
+      httpRequestTimeout: widget.httpRequestTimeout,
+      compoundClause: widget.compoundClause,
     );
   }
 }
@@ -1221,6 +1223,22 @@ class ReactiveGoogleMap extends StatefulWidget {
   /// ```
   final Map? distinctFieldConfig;
 
+  /// Configure whether the DSL query is generated with the compound clause of [CompoundClauseType.must] or [CompoundClauseType.filter]. If nothing is passed the default is to use [CompoundClauseType.must].
+  /// Setting the compound clause to filter allows search engine to cache and allows for higher throughput in cases where scoring isn’t relevant (e.g. term, geo or range type of queries that act as filters on the data)
+  ///
+  /// This property only has an effect when the search engine is either elasticsearch or opensearch.
+  /// > Note: `compoundClause` is supported with v8.16.0 (server) as well as with serverless search.
+  ///
+  ///   ///
+  /// For example,
+  /// ```dart
+  /// SearchBox(
+  ///   ...
+  ///   compoundClause:  CompoundClauseType.filter
+  /// )
+  /// ```
+  CompoundClauseType? compoundClause;
+
   /* ---- callbacks to create the side effects while querying ----- */
 
   /// It is a callback function which accepts component's future **value** as a
@@ -1531,7 +1549,11 @@ class ReactiveGoogleMap extends StatefulWidget {
   /// Zoom level to stop cluster rendering
   final double? stopClusteringZoom;
 
-  const ReactiveGoogleMap({
+  /// This prop is used to set the timeout value for HTTP requests.
+  /// Defaults to 30 seconds.
+  final Duration httpRequestTimeout;
+
+  ReactiveGoogleMap({
     Key? key,
     required this.id,
     this.showMarkerClusters = true,
@@ -1601,6 +1623,8 @@ class ReactiveGoogleMap extends StatefulWidget {
     this.results,
     this.distinctField,
     this.distinctFieldConfig,
+    this.httpRequestTimeout = const Duration(seconds: 30),
+    this.compoundClause,
 
     // Google map props
     required this.initialCameraPosition,