+++ searchbase/lib/src/searchcontroller.dart
@@ -1,4 +1,6 @@
+import 'dart:async';
 import 'dart:convert';
+import 'dart:io';
 import 'package:http/http.dart' as http;
 import 'base.dart';
 import 'searchbase.dart';
@@ -27,6 +29,13 @@ const suggestionQueryID = 'DataSearch__suggestions';
 /// -   a location filter widget,
 /// -   a widget to render the search results.
 class SearchController extends Base {
+  http.Client? _httpClient = null;
+
+  http.Client get httpClient {
+    _httpClient ??= http.Client();
+    return _httpClient!;
+  }
+
   // RS API properties
 
   /// A unique identifier of the component, can be referenced in other widgets' `react` prop to reactively update data.
@@ -129,7 +138,7 @@ class SearchController extends Base {
   /// To set the number of buckets to be returned by aggregations.
   ///
   /// > Note: This is a new feature and only available for appbase versions >= 7.41.0.
-  final int? aggregationSize;
+  int? aggregationSize;
 
   /// This property can be used to implement the pagination for `aggregations`.
   ///
@@ -345,7 +354,7 @@ class SearchController extends Base {
   /// This prop returns only the distinct value documents for the specified field.
   /// It is equivalent to the DISTINCT clause in SQL. It internally uses the collapse feature of Elasticsearch.
   /// You can read more about it over here - https://www.elastic.co/guide/en/elasticsearch/reference/current/collapse-search-results.html
-  final String? distinctField;
+  String? distinctField;
 
   /// This prop allows specifying additional options to the distinctField prop.
   /// Using the allowed DSL, one can specify how to return K distinct values (default value of K=1),
@@ -369,7 +378,27 @@ class SearchController extends Base {
   ///   'max_concurrent_group_searches': 4, },
   /// )
   /// ```
-  final Map? distinctFieldConfig;
+  Map? distinctFieldConfig;
+
+  /// This prop is used to set the timeout value for HTTP requests.
+  /// Defaults to 30 seconds.
+  Duration httpRequestTimeout;
+
+  /// Configure whether the DSL query is generated with the compound clause of [CompoundClauseType.must] or [CompoundClauseType.filter]. If nothing is passed the default is to use [CompoundClauseType.must].
+  /// Setting the compound clause to filter allows search engine to cache and allows for higher throughput in cases where scoring isn’t relevant (e.g. term, geo or range type of queries that act as filters on the data)
+  ///
+  /// This property only has an effect when the search engine is either elasticsearch or opensearch.
+  /// > Note: `compoundClause` is supported with v8.16.0 (server) as well as with serverless search.
+  ///
+  ///   ///
+  /// For example,
+  /// ```dart
+  /// SearchBox(
+  ///   ...
+  ///   compoundClause:  CompoundClauseType.filter
+  /// )
+  /// ```
+  CompoundClauseType? compoundClause;
 
   /* ---- callbacks to create the side effects while querying ----- */
 
@@ -390,7 +419,7 @@ class SearchController extends Base {
   ///   // or Future.error()
   /// }
   /// ```
-  final Future Function(dynamic value)? beforeValueChange;
+  Future Function(dynamic value)? beforeValueChange;
 
   /* ------------- change events -------------------------------- */
 
@@ -398,26 +427,25 @@ class SearchController extends Base {
   ///
   /// This property is handy in cases where you want to generate a side-effect on value selection.
   /// For example: You want to show a pop-up modal with the valid discount coupon code when a user searches for a product in a [SearchController].
-  final void Function(dynamic next, {dynamic prev})? onValueChange;
+  void Function(dynamic next, {dynamic prev})? onValueChange;
 
   /// It can be used to listen for the `results` changes.
-  final void Function(Results next, {Results prev})? onResults;
+  void Function(Results next, {Results prev})? onResults;
 
   /// It can be used to listen for the `aggregationData` property changes.
-  final void Function(Aggregations next, {Aggregations prev})?
-      onAggregationData;
+  void Function(Aggregations next, {Aggregations prev})? onAggregationData;
 
   /// It gets triggered in case of an error while fetching results.
-  final void Function(dynamic error)? onError;
+  void Function(dynamic error)? onError;
 
   /// It can be used to listen for the request status changes.
-  final void Function(String? next, {String? prev})? onRequestStatusChange;
+  void Function(String? next, {String? prev})? onRequestStatusChange;
 
   /// It is a callback function which accepts widget's **nextQuery** and **prevQuery** as parameters.
   ///
   /// It is called everytime the widget's query changes.
   /// This property is handy in cases where you want to generate a side-effect whenever the widget's query would change.
-  final void Function(List<Map>? next, {List<Map>? prev})? onQueryChange;
+  void Function(List<Map>? next, {List<Map>? prev})? onQueryChange;
 
   /* ------ Private properties only for the internal use ----------- */
   SearchBase? _parent;
@@ -488,6 +516,8 @@ class SearchController extends Base {
     this.distinctFieldConfig,
     this.value,
     this.clearOnQueryChange = false,
+    this.httpRequestTimeout = const Duration(seconds: 30),
+    this.compoundClause,
   }) : super(index, url, credentials,
             appbaseConfig: appbaseConfig,
             transformRequest: transformRequest,
@@ -588,6 +618,7 @@ class SearchController extends Base {
       'index': this.index,
       'distinctField': distinctField,
       'distinctFieldConfig': distinctFieldConfig,
+      'compoundClause': compoundClause,
     };
     query.removeWhere((key, value) => key == null || value == null);
     return query;
@@ -628,15 +659,19 @@ class SearchController extends Base {
 
   /// can be used to set the `value` property
   void setValue(dynamic value, {Options? options}) async {
-    if (this.beforeValueChange != null) {
-      try {
+    try {
+      if (this.beforeValueChange != null) {
         var val = await beforeValueChange!(value);
         this._performUpdate(val, options);
-      } catch (e) {
-        print(e);
+      } else {
+        this._performUpdate(value, options);
       }
-    } else {
-      this._performUpdate(value, options);
+
+      // Check if options contain a completer and complete it if it exists
+      options?.completer?.complete();
+    } catch (e) {
+      // If any error occurs, throw an exception for a failed operation
+      throw Exception('Error in setValue: $e');
     }
   }
 
@@ -759,8 +794,11 @@ class SearchController extends Base {
     try {
       this.updateQuery();
       var searchbaseInstance = this._parent;
-      var shouldAddToWaitList = searchbaseInstance!
-          .shouldAddRequestToWaitList(this.id, true, this.query);
+      var shouldAddToWaitList = searchbaseInstance!.shouldAddRequestToWaitList(
+        this.id,
+        true,
+        this.query,
+      );
       if (!shouldAddToWaitList) {
         this._lastUsedDefaultQueryTimeStamp = currentTimeStamp;
         this.setRequestStatus(RequestStatus.PENDING, options: options);
@@ -1022,7 +1060,7 @@ class SearchController extends Base {
     final String url =
         "${this.url}/_analytics/${this._getSearchIndex(false)}/recent-searches?${queryString}";
     try {
-      final res = await http.get(Uri.parse(url), headers: this.headers);
+      final res = await httpClient.get(Uri.parse(url), headers: this.headers);
       if (res.statusCode >= 500) {
         return Future.error(res);
       }
@@ -1151,6 +1189,100 @@ class SearchController extends Base {
     };
   }
 
+  void updateConfig(Map searchController) {
+    // Update the componentConfig with the provided values if they are not null
+    this.index = searchController["index"] ?? this.index;
+    this.url = searchController["url"] ?? this.url;
+    this.credentials = searchController["credentials"] ?? this.credentials;
+    this.headers =
+        searchController["headers"] as Map<String, String>? ?? this.headers;
+    this.transformRequest =
+        searchController["transformRequest"] as TransformRequest? ??
+            this.transformRequest;
+    this.transformResponse =
+        searchController["transformResponse"] as TransformResponse? ??
+            this.transformResponse;
+    this.appbaseConfig =
+        searchController["appbaseConfig"] as AppbaseSettings? ??
+            this.appbaseConfig;
+    this.type = searchController["type"] ?? this.type;
+    this.dataField = searchController["dataField"] ?? this.dataField;
+    this.react = searchController["react"] ?? this.react;
+    this.queryFormat = searchController["queryFormat"] ?? this.queryFormat;
+    this.categoryField =
+        searchController["categoryField"] ?? this.categoryField;
+    this.categoryValue =
+        searchController["categoryValue"] ?? this.categoryValue;
+    this.nestedField = searchController["nestedField"] ?? this.nestedField;
+    this.from = searchController["from"] ?? this.from;
+    this.size = searchController["size"] ?? this.size;
+    this.sortBy = searchController["sortBy"] ?? this.sortBy;
+    this.aggregationField =
+        searchController["aggregationField"] ?? this.aggregationField;
+    this.aggregationSize =
+        searchController["aggregationSize"] ?? this.aggregationSize;
+    this.after = searchController["after"] ?? this.after;
+    this.includeNullValues =
+        searchController["includeNullValues"] ?? this.includeNullValues;
+    this.includeFields =
+        searchController["includeFields"] ?? this.includeFields;
+    this.excludeFields =
+        searchController["excludeFields"] ?? this.excludeFields;
+    this.fuzziness = searchController["fuzziness"] ?? this.fuzziness;
+    this.searchOperators =
+        searchController["searchOperators"] ?? this.searchOperators;
+    this.highlight = searchController["highlight"] ?? this.highlight;
+    this.highlightField =
+        searchController["highlightField"] ?? this.highlightField;
+    this.customHighlight =
+        searchController["customHighlight"] ?? this.customHighlight;
+    this.interval = searchController["interval"] ?? this.interval;
+    this.aggregations = searchController["aggregations"] ?? this.aggregations;
+    this.missingLabel = searchController["missingLabel"] ?? this.missingLabel;
+    this.showMissing = searchController["showMissing"] ?? this.showMissing;
+    this.enableSynonyms =
+        searchController["enableSynonyms"] ?? this.enableSynonyms;
+    this.selectAllLabel =
+        searchController["selectAllLabel"] ?? this.selectAllLabel;
+    this.pagination = searchController["pagination"] ?? this.pagination;
+    this.queryString = searchController["queryString"] ?? this.queryString;
+    this.defaultQuery = searchController["defaultQuery"] ?? this.defaultQuery;
+    this.customQuery = searchController["customQuery"] ?? this.customQuery;
+    this.beforeValueChange =
+        searchController["beforeValueChange"] ?? this.beforeValueChange;
+    this.onValueChange =
+        searchController["onValueChange"] ?? this.onValueChange;
+    this.onResults = searchController["onResults"] ?? this.onResults;
+    this.onAggregationData =
+        searchController["onAggregationData"] ?? this.onAggregationData;
+    this.onError = searchController["onError"] ?? this.onError;
+    this.onRequestStatusChange =
+        searchController["onRequestStatusChange"] ?? this.onRequestStatusChange;
+    this.onQueryChange =
+        searchController["onQueryChange"] ?? this.onQueryChange;
+    this.enablePopularSuggestions =
+        searchController["enablePopularSuggestions"] ??
+            this.enablePopularSuggestions;
+    this.maxPopularSuggestions =
+        searchController["maxPopularSuggestions"] ?? this.maxPopularSuggestions;
+    this.showDistinctSuggestions =
+        searchController["showDistinctSuggestions"] ??
+            this.showDistinctSuggestions;
+    this.preserveResults =
+        searchController["preserveResults"] ?? this.preserveResults;
+    this.clearOnQueryChange =
+        searchController["clearOnQueryChange"] ?? this.clearOnQueryChange;
+    this.value = searchController["value"] ?? this.value;
+    this.distinctField =
+        searchController["distinctField"] ?? this.distinctField;
+    this.distinctFieldConfig =
+        searchController["distinctFieldConfig"] ?? this.distinctFieldConfig;
+    this.compoundClause =
+        searchController["compoundClause"] ?? this.compoundClause;
+    this.httpRequestTimeout =
+        searchController['httpRequestTimeout'] ?? this.httpRequestTimeout;
+  }
+
   // Method to apply the changes based on set options
   void applyOptions(
       Options? options, KeysToSubscribe key, prevValue, nextValue) {
@@ -1187,12 +1319,15 @@ class SearchController extends Base {
   }
 
   Future<Map> _fetchRequest(
-      Map requestBody, bool isPopularSuggestionsAPI) async {
+    Map requestBody,
+    bool isPopularSuggestionsAPI,
+  ) async {
     // remove undefined properties from request body
     final requestOptions = {
       'body': jsonEncode(requestBody),
-      'headers': {...?this.headers}
+      'headers': {...?this.headers},
     };
+
     try {
       final finalRequestOptions =
           await this._handleTransformRequest(requestOptions);
@@ -1201,11 +1336,19 @@ class SearchController extends Base {
       String suffix = '_reactivesearch.v3';
       final String url =
           "${this.url}/${this._getSearchIndex(isPopularSuggestionsAPI)}/$suffix";
-      final http.Response res = await http.post(
+
+      final timeoutDuration = httpRequestTimeout;
+
+      final http.Response res = await httpClient
+          .post(
         Uri.parse(url),
         headers: finalRequestOptions['headers'],
         body: finalRequestOptions['body'],
-      );
+      )
+          .timeout(timeoutDuration, onTimeout: () {
+        throw TimeoutException('Request took too long to complete');
+      });
+
       final responseHeaders = res.headers;
       // check if search component is present
       final queryID = responseHeaders['x-search-id'];
@@ -1218,22 +1361,28 @@ class SearchController extends Base {
         }
       }
       if (res.statusCode >= 500) {
-        return Future.error(res);
+        throw HttpError(res.statusCode, 'Internal Server Error');
       }
       if (res.statusCode >= 400) {
-        final data = jsonDecode(res.body);
-        return Future.error(data);
+        final errorResponse = jsonDecode(res.body);
+        final errorStatusCode = res.statusCode;
+        final errorMessage = errorResponse['message'] ?? 'Unknown Error';
+        throw HttpError(errorStatusCode, errorMessage);
       }
       final data = jsonDecode(res.body);
       final transformedData = await this._handleTransformResponse(data);
       if (transformedData.containsKey('error')) {
-        return Future.error(transformedData);
+        throw transformedData; // Throw transformed error response
       }
-      return Future.value({
+      return {
         ...transformedData,
         '_timestamp': timestamp,
-        '_headers': responseHeaders
-      });
+        '_headers': responseHeaders,
+      };
+    } on TimeoutException {
+      throw HttpTimeout(); // Throw HttpTimeout error
+    } on SocketException {
+      throw NoInternet(); // Throw NoInternet error
     } catch (e) {
       print(e);
       return Future.error(e);