+++ searchbase/lib/src/types.dart
@@ -2,6 +2,7 @@ import 'results.dart';
 import 'aggregations.dart';
 import 'constants.dart';
 import 'searchcontroller.dart';
+import 'dart:async';
 
 typedef TransformRequest = Future<Object> Function(Map requestOptions);
 typedef TransformResponse = Future Function(dynamic response);
@@ -245,10 +246,14 @@ class Options {
   bool? triggerDefaultQuery;
   bool? triggerCustomQuery;
   bool? stateChanges;
-  Options(
-      {this.triggerDefaultQuery,
-      bool? triggerCustomQuery,
-      bool? stateChanges}) {
+  Completer<void>? completer;
+
+  Options({
+    this.triggerDefaultQuery,
+    bool? triggerCustomQuery,
+    bool? stateChanges,
+    this.completer,
+  }) {
     this.triggerDefaultQuery =
         triggerDefaultQuery != null ? triggerDefaultQuery : false;
     this.triggerCustomQuery =
@@ -506,5 +511,45 @@ extension KeysToSubscribeExtension on KeysToSubscribe {
       default:
         return "";
     }
+  }
+}
+
+/// Exception class representing the absence of an internet connection.
+class NoInternet implements Exception {
+  final String message;
+
+  /// Creates a [NoInternet] exception with an optional [message].
+  NoInternet([this.message = 'No internet connection']);
+
+  @override
+  String toString() {
+    return 'NoInternet: $message';
+  }
+}
+
+/// Exception class representing an HTTP request timeout.
+class HttpTimeout implements Exception {
+  final String message;
+
+  /// Creates an [HttpTimeout] exception with an optional [message].
+  HttpTimeout([this.message = 'HTTP request timed out']);
+
+  @override
+  String toString() {
+    return 'HttpTimeout: $message';
+  }
+}
+
+/// Exception class representing an HTTP error with an associated status code and an optional message.
+class HttpError implements Exception {
+  final int statusCode;
+  final String? message;
+
+  /// Creates an [HttpError] exception with the given [statusCode] and an optional [message].
+  HttpError(this.statusCode, [this.message]);
+
+  @override
+  String toString() {
+    return 'HttpError: $statusCode${message != null ? " ($message)" : ""}';
   }
 }